DatabaseManager changes

+++ Labanoro_Draugai/Labanoro_Draugai/src/main/java/lt/vu/mif/labanoro_draugai/entities/Service.java
@@ -6,6 +6,7 @@
 package lt.vu.mif.labanoro_draugai.entities;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import javax.persistence.Basic;
@@ -71,8 +72,9 @@ public class Service implements Serializable {
     @Column(name = "SEASONENDDATE")
     @Temporal(TemporalType.DATE)
     private Date seasonenddate;
+    // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
     @Column(name = "WEEKPRICE")
-    private Integer weekprice;
+    private BigDecimal weekprice;
     @Column(name = "NUMBEROFPLACES")
     private Integer numberofplaces;
     @Column(name = "ISDELETED")
@@ -155,11 +157,11 @@ public class Service implements Serializable {
         this.seasonenddate = seasonenddate;
     }
 
-    public Integer getWeekprice() {
+    public BigDecimal getWeekprice() {
         return weekprice;
     }
 
-    public void setWeekprice(Integer weekprice) {
+    public void setWeekprice(BigDecimal weekprice) {
         this.weekprice = weekprice;
     }
